apply reviews in testcode
